Refactored SubMeta to inline links
